Idle mode instead of shutting down
Added sleep mode, originally kiisu shuts down after the fake backlight turns off (usually 30s, adjustable in settings). This change makes it instead go in a idle mode. MCU is still active, but the OLED turns off to save battery and prevent burn in. After 30 minutes of idling, it will shut down. The period for it to start idling is adjustable in settings>lcd>backlight time.